requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3
PyYAML==6.0.1
pygame==2.5.2
pandas==2.0.3
tensorflow==2.15.0
scikit-learn==1.3.2
matplotlib==3.7.3
seaborn==0.12.2
